requirements.txt
django
psycopg2-binary
python-dotenv
djangorestframework
drf-spectacular

djangorestframework-simplejwt
django-storages[azure]
django-cors-headers


